requirements_test.txt
cov-core==1.7
coverage==3.7.1
nose==1.3.0
pep8==1.4.6
pyflakes==0.7.3
pylint==1.1.0
pytest-cache==1.0
pytest-cov==1.6
pytest-django==2.6
pytest-pep8==1.0.5
pytest==2.5.2
